ATS System/requirements.txt
gradio
google-generativeai
python-dotenv
pdf2image
